refactor: Rewrite the FSM implementation
This new implementation uses XState and does not duplicate FSM logic to
handle dead run or historical messages.

The API in regards to the environment variables has not changed.

Priority queues are no longer needed assuming that the event timestamps
have been given in monotonic order by the previous components in the
pipeline.

Compared to the previous implementation, this one might use more memory
and CPU due to the many features in XState. However, the code seems
cleaner and we've managed to drop the line count.

Functions isDeparture and useCurrent are tricky for they need to handle
many HFP situations.

To match Transmodel, internally the name "dead run" is used. However,
not to break backwards compatibility, the environment variables still
use the previous format "DEADRUN".