test: Fix tests
Unfortunately, pulsar-client-node does not seem very dependable and the
data-driven integration tests would often fail flakily probably due to
a bug in pulsar-client-node. We use a workaround of syncing each message
separately to the Pulsar cluster for the test producer inputs.

Some of the data-driven tests required more HFP context than before so
that the state machine would realize that those examples are not from the
beginnings of service journeys.