fix: Discard HFP messages for metro
HFP messages for metro do not support PDE or DEP events which our
current implementation uses to recognize departures from stops. If the
relevant parts of vehicleActor.ts are changed to recognize departures
from the unreliable next stop data in the HFP MQTT topic, metro messages
can be used again.

If the metro messages were not to be discarded, the current logic would
never acknowledge them back to Pulsar. That would cause a space leak and
force every new instance of this service to trawl through the metro message
history.